feat(v2): update ludpo to v4.0.0
