--- .eslintrc.yml
@@ -1,34 +1,0 @@
-env:
-  browser: true
-  commonjs: true
-  es6: true
-parser: "@typescript-eslint/parser"
-plugins:
-  - "prettier"
-extends:
-  - "plugin:@typescript-eslint/recommended"
-  - "plugin:prettier/recommended"
-globals:
-  Atomics: readonly
-  SharedArrayBuffer: readonly
-parserOptions:
-  ecmaVersion: 2018
-rules:
-  "@typescript-eslint/explicit-function-return-type":
-    - warn
-    - allowExpressions: true
-  "@typescript-eslint/explicit-member-accessibility": off
-  "@typescript-eslint/ban-types": off
-  "@typescript-eslint/no-angle-bracket-type-assertion": off
-  "@typescript-eslint/no-explicit-any": off
-  "@typescript-eslint/no-non-null-assertion": off
-  "@typescript-eslint/no-parameter-properties": off
-  "@typescript-eslint/prefer-interface": off
-  "@typescript-eslint/indent": off
-overrides:
-  - files:
-      - "*.test.ts"
-      - "*.spec.ts"
-    rules:
-      "@typescript-eslint/no-empty-function": off
-      "@typescript-eslint/interface-name-prefix": off
