+++ .github/workflows/build.yml
@@ -7,35 +7,60 @@ on:
 
 jobs:
   build:
+    name: Run build with ${{ matrix.node-version }}
     runs-on: ubuntu-latest
-
     strategy:
-      fail-fast: false
       matrix:
-        node-version: [16.x, 18.x]
-        database-type: [sqlite3]
-
+        node-version: [22.x, 20.x, 18.x]
     steps:
-      - uses: actions/checkout@v3
+      - name: Checkout code
+        uses: actions/checkout@v3
 
-      - name: Use Node.js ${{ matrix.node-version }}
+      - name: Setup Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
 
-      - name: Start Database
-        run: sqlite3 .${{ matrix.database-type }}.db < src/__tests__/data/sqlite.sql
-
-      - name: Run npm install
+      - name: Install dependencies
         run: npm ci
+        env:
+          CI: true
+
+      - name: Lint code
+        run: npm run lint
+        env:
+          CI: true
+
+      - name: Unit test
+        run: npm run test:unit
+        env:
+          CI: true
+
+  integration:
+    name: Run integration tests
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: 22.x
+
+      - name: Install dependencies
+        run: npm ci
+        env:
+          CI: true
+
+      - name: Start Database
+        run: sqlite3 .sqlite3.db < .devcontainer/init/sqlite.sql
 
       - name: Build and test
         run: |
-          npm run build
-          npm run test:db
+          npm run test:integration
         env:
           CI: true
-          DB: ${{ matrix.database-type }}
 
-      - name: Stop Database
-        run: rm .${{ matrix.database-type }}.db
+      - name: Stop Database(s)
+        run: rm .sqlite3.db
