+++ .prettierrc
@@ -1,12 +1,16 @@
 {
-    "printWidth": 120,
-    "tabWidth": 4,
-    "useTabs": false,
+    "printWidth": 80,
+    "tabWidth": 2,
     "semi": true,
     "singleQuote": true,
     "quoteProps": "as-needed",
-    "trailingComma": "none",
+    "trailingComma": "all",
     "bracketSpacing": true,
-    "arrowParens": "avoid",
-    "endOfLine": "lf"
+    "bracketSameLine": false,
+    "arrowParens": "always",
+    "proseWrap": "preserve",
+    "vueIndentScriptAndStyle": true,
+    "endOfLine": "lf",
+    "embeddedLanguageFormatting": "auto",
+    "singleAttributePerLine": false
 }
