--- test/jest.setup.ts
@@ -1,1 +1,0 @@
-import '../src/';
