--- typescript/tsconfig.base.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "removeComments": true,
-        "strict": true,
-        "noImplicitAny": true,
-        "strictNullChecks": true,
-        "alwaysStrict": true,
-        "noImplicitThis": true,
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "noImplicitReturns": true,
-        "allowSyntheticDefaultImports": true,
-        "experimentalDecorators": true,
-        "emitDecoratorMetadata": true,
-        "outDir": "dist",
-        "stripInternal": true,
-        "moduleResolution": "node",
-        "importHelpers": true
-    },
-    "include": ["../src/"],
-    "exclude": ["../src/__tests__"]
-}
