+++ typescript/tsconfig.esm.json
@@ -1,10 +1,23 @@
 {
-    "extends": "./tsconfig.base.json",
+    "extends": "../tsconfig.json",
     "compilerOptions": {
+        "moduleResolution": "node",
+        "declaration": false,
+        "sourceMap": false,
+        "incremental": false,
         "target": "ES2020",
         "module": "ES2020",
         "outDir": "../dist/esm",
-        "lib": ["es2020"],
+        "lib": [
+            "ES2020",
+            "DOM"
+        ],
         "downlevelIteration": true
-    }
+    },
+    "include": [
+        "../src/"
+    ],
+    "exclude": [
+        "../**/*.test.ts"
+    ]
 }
